requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.11.7
python-dotenv>=1.0.0
requests>=2.31.0
